+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/di/KeyEventHandlerImpl.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.window.di
+
+import androidx.compose.ui.input.key.KeyEvent
+
+//TODO maybe redundant and easier to implement KeyEventHandler interface in InputServiceImpl
+internal class KeyEventHandlerImpl(
+    val inputService: InputServiceImpl,
+):KeyEventHandler {
+    override fun onPreviewKeyEvent(event: KeyEvent): Boolean {
+        return inputService.onPreviewKeyEvent(event)
+    }
+
+    override fun onKeyEvent(event: KeyEvent): Boolean {
+        // Do nothing
+        return false
+    }
+}
