+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/di/KeyboardEventHandlerImpl.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.window.di
+
+import androidx.compose.ui.ComposeScene
+import androidx.compose.ui.input.key.KeyEvent
+import org.jetbrains.skiko.SkikoKeyboardEvent
+
+class KeyboardEventHandlerImpl(
+    val sceneProvider: () -> ComposeScene,
+) : KeyboardEventHandler {
+    val scene get() = sceneProvider()
+    override fun onKeyboardEvent(event: SkikoKeyboardEvent) {
+        scene.sendKeyEvent(KeyEvent(event))
+    }
+}
