+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/di/SkikoUIViewDelegate.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.window.di
+
+import androidx.compose.ui.interop.UIKitInteropTransaction
+import androidx.compose.ui.window.UITouchesEventPhase
+import kotlinx.cinterop.CValue
+import org.jetbrains.skia.Canvas
+import platform.CoreGraphics.CGPoint
+import platform.Foundation.NSTimeInterval
+import platform.UIKit.UIEvent
+import platform.UIKit.UIView
+
+internal interface SkikoUIViewDelegate {
+
+    fun pointInside(point: CValue<CGPoint>, event: UIEvent?): Boolean
+
+    fun onTouchesEvent(view: UIView, event: UIEvent, phase: UITouchesEventPhase)
+
+    fun retrieveInteropTransaction(): UIKitInteropTransaction
+
+    fun render(canvas: Canvas, targetTimestamp: NSTimeInterval)
+
+    /**
+     * A callback invoked when [UIView.didMoveToWindow] receives non null window
+     */
+    fun onAttachedToWindow() {}
+}
